add order select hook support (#109)
* add order select hook support

* linting

* fix import

* change subproj commit

* log from context
