+++ src/server.ts
@@ -3,7 +3,8 @@ import bodyParser from 'body-parser';
 import container from './lib/winston';
 import morgan from 'morgan';
 import { CdsService } from './rems-cds-hooks/resources/CdsService';
-import remsService from './hooks/rems.hook';
+import remsService from './hooks/rems.ordersign';
+import orderSelectService from './hooks/rems.orderselect';
 import { Server } from '@projecttacoma/node-fhir-server-core';
 import Etasu from './lib/etasu';
 import env from 'env-var';
@@ -96,7 +97,7 @@ class REMSServer extends Server {
   registerCdsHooks({ discoveryEndpoint = '/cds-services' }: any) {
     this.cdsHooksEndpoint = discoveryEndpoint;
     this.registerService(remsService);
-
+    this.registerService(orderSelectService);
     this.app.get(discoveryEndpoint, (req: any, res: { json: (arg0: { services: any }) => any }) =>
       res.json({ services: this.services })
     );
